pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "security-incident-soc"
version = "1.0.0"
description = "Deterministic SOC incident response simulation for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "pydantic>=2.8.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "openai>=1.0.0",
]

[project.scripts]
server = "app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "graders*", "tasks*", "server*"]
exclude = ["venv*", "*.pyc"]